+++ pymeasure/instruments/instrument.py
@@ -134,6 +134,14 @@ class Instrument(CommonBase):
         else:
             raise NotImplementedError("Non SCPI instruments require implementation in subclasses")
 
+    @property
+    def next_error(self):
+        """Get the next error of the instrument (tuple of code and message)."""
+        if self.SCPI:
+            return self.values("SYST:ERR?")
+        else:
+            raise NotImplementedError("Non SCPI instruments require implementation in subclasses")
+
     # Wrapper functions for the Adapter object
     def write(self, command, **kwargs):
         """Write a string command to the instrument appending `write_termination`.
@@ -212,7 +220,7 @@ class Instrument(CommonBase):
         if self.SCPI:
             errors = []
             while True:
-                err = self.values("SYST:ERR?")
+                err = self.next_error
                 if int(err[0]) != 0:
                     log.error(f"{self.name}: {err[0]}, {err[1]}")
                     errors.append(err)