+++ pymeasure/instruments/keithley/keithley2260B.py
@@ -129,7 +129,7 @@ class Keithley2260B(Instrument):
         self.output_enabled = value
 
     @property
-    def error(self):
+    def next_error(self):
         """ Returns a tuple of an error code and message from a
         single error. """
         err = self.values(":system:error?")
@@ -139,14 +139,16 @@ class Keithley2260B(Instrument):
         message = err[1].replace('"', "")
         return (code, message)
 
+    error = next_error
+
     def check_errors(self):
         """ Logs any system errors reported by the instrument.
         """
-        code, message = self.error
+        code, message = self.next_error
         while code != 0:
             t = time.time()
             log.info("Keithley 2260B reported error: %d, %s" % (code, message))
-            code, message = self.error
+            code, message = self.next_error
             if (time.time() - t) > 10:
                 log.warning("Timed out for Keithley 2260B error retrieval.")