+++ pymeasure/instruments/keithley/keithley2400.py
@@ -24,6 +24,7 @@
 
 import logging
 import time
+from warnings import warn
 
 import numpy as np
 
@@ -508,7 +509,7 @@ class Keithley2400(KeithleyBuffer, Instrument):
     )
 
     @property
-    def error(self):
+    def next_error(self):
         """ Returns a tuple of an error code and message from a
         single error. """
         err = self.values(":system:error?")
@@ -518,14 +519,19 @@ class Keithley2400(KeithleyBuffer, Instrument):
         message = err[1].replace('"', '')
         return (code, message)
 
+    @property
+    def error(self):
+        warn("Deprecated to use `error`, use `next_error` instead.", FutureWarning)
+        return self.next_error
+
     def check_errors(self):
         """ Logs any system errors reported by the instrument.
         """
-        code, message = self.error
+        code, message = self.next_error
         while code != 0:
             t = time.time()
             log.info("Keithley 2400 reported error: %d, %s" % (code, message))
-            code, message = self.error
+            code, message = self.next_error
             if (time.time() - t) > 10:
                 log.warning("Timed out for Keithley 2400 error retrieval.")