+++ pymeasure/instruments/keithley/keithley2450.py
@@ -24,6 +24,7 @@
 
 import logging
 import time
+from warnings import warn
 
 import numpy as np
 
@@ -495,7 +496,7 @@ class Keithley2450(KeithleyBuffer, Instrument):
         self.beep(base_frequency * 6.0 / 4.0, duration)
 
     @property
-    def error(self):
+    def next_error(self):
         """ Returns a tuple of an error code and message from a
         single error. """
         err = self.values(":system:error?")
@@ -505,14 +506,19 @@ class Keithley2450(KeithleyBuffer, Instrument):
         message = err[1].replace('"', '')
         return (code, message)
 
+    @property
+    def error(self):
+        warn("Deprecated to use `error`, use `next_error` instead.", FutureWarning)
+        return self.next_error
+
     def check_errors(self):
         """ Logs any system errors reported by the instrument.
         """
-        code, message = self.error
+        code, message = self.next_error
         while code != 0:
             t = time.time()
             log.info("Keithley 2450 reported error: %d, %s", code, message)
-            code, message = self.error
+            code, message = self.next_error
             if (time.time() - t) > 10:
                 log.warning("Timed out for Keithley 2450 error retrieval.")